Fix lexer block style, move back to main folder
